pyproject.toml
[project]
name = "buaa-evaluation"
version = "1.1.0"
description = "Teaching evaluation assistant for BUAA"
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }
authors = [
    { name = "ZenAlexa" }
]
keywords = ["buaa", "education", "evaluation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Desktop Environment",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "requests>=2.28.0",
    "beautifulsoup4>=4.12.0",
    "pywebview>=5.0",
]

[project.optional-dependencies]
dev = [
    "pyinstaller>=6.0",
]
macos = [
    "pyobjc>=10.0",
]

[project.urls]
Homepage = "https://github.com/ZenAlexa/BUAA_TeachingEvaluation_2024Winter"
Repository = "https://github.com/ZenAlexa/BUAA_TeachingEvaluation_2024Winter"
Issues = "https://github.com/ZenAlexa/BUAA_TeachingEvaluation_2024Winter/issues"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["backend*"]

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N"]
ignore = ["E501"]
